Integrated Alembic for database migrations and implemented relational mapping between users and tasks

+++ src/tasks/controller.py
@@ -2,13 +2,14 @@ from src.tasks.dtos import TaskSchema
 from sqlalchemy.orm import Session
 from src.tasks.models import TaskModel
 from fastapi import HTTPException
-
+from src.user.models import UserModel
 #post/create
-def create_task(body:TaskSchema,db:Session):
+def create_task(body:TaskSchema,db:Session,user:UserModel):
     data=body.model_dump()
     new_task=TaskModel(title=data["title"],
                        description=data["description"],
-                       is_completed=data["is_completed"]
+                       is_completed=data["is_completed"],
+                       user_id=user.id
                         )
     db.add(new_task)
     db.commit()
@@ -18,8 +19,8 @@ def create_task(body:TaskSchema,db:Session):
 
 #get
 
-def get_tasks(db:Session):
-    tasks=db.query(TaskModel).all()
+def get_tasks(db:Session,user:UserModel):
+    tasks=db.query(TaskModel).filter(TaskModel.user_id == user.id).all()
     return tasks
 
 
@@ -36,11 +37,13 @@ def get_one_task(task_id: int,db:Session):
 #update/put
 
 
-def update_task(body:TaskSchema,task_id:int,db:Session):
+def update_task(body:TaskSchema,task_id:int,db:Session,user:UserModel):
     one_task=db.query(TaskModel).get(task_id)
     if not one_task:
         raise HTTPException(404,detail="task id is incorrect")
     
+    if one_task.user_id!=user.id:
+        raise HTTPException(401,detail="You are not allowed to update this task")
     #if we want to update mutiple or one at a time 
     body=body.model_dump()
     for field,value in body.items():
@@ -58,11 +61,12 @@ def update_task(body:TaskSchema,task_id:int,db:Session):
 
 
 
-def delete_task(task_id: int,db:Session):
+def delete_task(task_id: int,db:Session,user:UserModel):
     one_task=db.query(TaskModel).get(task_id)
     if not one_task:
         raise HTTPException(404,detail="task id is incorrect")
-    
+    if one_task.user_id!=user.id:
+        raise HTTPException(401,detail="You are not to delete this task")
     db.delete(one_task)
     db.commit()
